refactor: move Register to pages, extract validation service
