+++ src/App.js
@@ -1,4 +1,4 @@
-import Register from "./components/Register";
+import RegisterPage from "./pages/RegisterPage";
 import Header from "./components/Header";
 import Footer from "./components/Footer";
 import { useTranslation } from "react-i18next";
@@ -13,7 +13,7 @@ export default function App() {
 
       <main className="flex-1 max-w-screen-xl mx-auto px-4 py-10">
         {path === "/register" ? (
-          <Register />
+          <RegisterPage />
         ) : (
           <>
             <h2 className="text-2xl font-semibold mb-4">{t("welcome")}</h2>
